backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
google-generativeai==0.7.2
gunicorn==22.0.0
numpy
pandas